Add required OSGi capability to wire o.e.swt to its native fragments
This ensures that the org.eclipse.swt bundle can only resolve if a the
native fragment of the current platform is available in the OSGi
runtime. This also has the effect that for launches, where dependencies
are added automatically, the native fragment of the current platform is
always added without anything external that enforces its presence (like
a feature).

Consequently this also allows to remove the p2 instructions doing the
same at p2 level.

Fixes https://github.com/eclipse-platform/eclipse.platform.swt/issues/490